add dockerfile

+++ movie-service/src/main/java/com/demo/cloud/entity/Movie.java
@@ -7,7 +7,7 @@ import lombok.Data;
 @Builder
 public class Movie {
 
-    private Long id;
+    private Float id;
 
     private String name;
 }
